re-arch numpy scalar/array
